requirements.txt
gunicorn
flask
pandas
matplotlib
tensorflow
numpy
seaborn
scikit-learn
opencv-python
torch
markupsafe
torchvision
Pillow
flask-mysqldb
passlib
bcrypt
requests
werkzeug
